feat: Added STS support.
