+++ packages/aws-sqs/src/_index.mts
@@ -1,4 +1,3 @@
+export * from "./adapter/_index.mjs";
 export * from "./definition/_index.mjs";
 export * from "./entity/_index.mjs";
-export * from "./predicate/_index.mjs";
-export * from "./service/_index.mjs";
