+++ packages/aws-sqs/src/adapter/_index.mts
@@ -1,0 +1,1 @@
+export * from "./sqs.adapter.mjs";
