+++ packages/aws-sqs/src/adapter/definition/interface/sqs-adapter-instantiation.interface.mts
@@ -1,0 +1,12 @@
+interface SQSAdapterInstantiationInterface
+{
+	accessKeyId: string;
+	accessSecret: string;
+	sessionToken?: string;
+	region: string;
+	host: string;
+	accountId: string;
+	https: boolean;
+}
+
+export type { SQSAdapterInstantiationInterface };
