--- packages/aws-sqs/src/definition/interface/_index.mts
@@ -1,7 +1,0 @@
-export * from "./message-attribute.interface.mjs";
-export * from "./message.interface.mjs";
-export * from "./receive-message-response.interface.mjs";
-export * from "./receive-message-result.interface.mjs";
-export * from "./response-envelope.interface.mjs";
-export * from "./response-metadata.interface.mjs";
-export * from "./sqs-service-instantiation.interface.mjs";
