--- packages/aws-sqs/src/definition/interface/receive-message-response.interface.mts
@@ -1,10 +1,0 @@
-import type { ReceiveMessageResultInterface } from "./receive-message-result.interface.mjs";
-import type { ResponseMetadataInterface } from "./response-metadata.interface.mjs";
-
-interface ReceiveMessageResponseInterface
-{
-	ReceiveMessageResult: ReceiveMessageResultInterface | null;
-	ResponseMetadata: ResponseMetadataInterface;
-}
-
-export type { ReceiveMessageResponseInterface };
