--- packages/aws-sqs/src/definition/interface/receive-message-result.interface.mts
@@ -1,9 +1,0 @@
-import type { MessageInterface } from "./message.interface.mjs";
-
-interface ReceiveMessageResultInterface
-{
-	messages?: Array<MessageInterface> | null;
-	Message?: MessageInterface | null;
-}
-
-export type { ReceiveMessageResultInterface };
