--- packages/aws-sqs/src/definition/interface/response-envelope.interface.mts
@@ -1,8 +1,0 @@
-import type { ReceiveMessageResponseInterface } from "./receive-message-response.interface.mjs";
-
-interface ResponseEnvelopeInterface
-{
-	ReceiveMessageResponse: ReceiveMessageResponseInterface;
-}
-
-export type { ResponseEnvelopeInterface };
