--- packages/aws-sqs/src/definition/interface/sqs-service-instantiation.interface.mts
@@ -1,11 +1,0 @@
-interface SQSServiceInstantiationInterface
-{
-	accessKeyId: string;
-	accessSecret: string;
-	region: string;
-	host: string;
-	accountId: string;
-	https: boolean;
-}
-
-export type { SQSServiceInstantiationInterface };
