+++ packages/aws-sqs/src/entity/_index.mts
@@ -1,5 +1,5 @@
-export * from "./message.mjs";
-export * from "./receive-message-response.mjs";
-export * from "./receive-message-result.mjs";
-export * from "./response-envelope.mjs";
-export * from "./response-metadata.mjs";
+export * from "./message/message.mjs";
+export * from "./receive-message-response/receive-message-response.mjs";
+export * from "./receive-message-result/receive-message-result.mjs";
+export * from "./response-envelope/response-envelope.mjs";
+export * from "./response-metadata/response-metadata.mjs";
