+++ packages/aws-sqs/src/entity/message-attribute/definition/interface/message-attribute.interface.mts
