+++ packages/aws-sqs/src/entity/message/definition/interface/message.interface.mts
@@ -1,0 +1,14 @@
+import type { MessageAttributeInterface } from "../../../message-attribute/definition/interface/message-attribute.interface.mjs";
+
+interface MessageInterface
+{
+	Attributes?: Array<MessageAttributeInterface> | null;
+	MessageId: string;
+	ReceiptHandle: string;
+	Body: string;
+	MD5OfBody: string;
+	MD5OfMessageAttributes?: string | null;
+	MessageAttributes?: Record<string, unknown> | null;
+}
+
+export type { MessageInterface };
