+++ packages/aws-sqs/src/entity/receive-message-response/receive-message-response.mts
@@ -1,0 +1,32 @@
+import type { ReceiveMessageResponseInterface } from "./definition/interface/receive-message-response.interface.mjs";
+
+import { ReceiveMessageResult } from "../receive-message-result/receive-message-result.mjs";
+import { ResponseMetadata } from "../response-metadata/response-metadata.mjs";
+
+class ReceiveMessageResponse
+{
+	private readonly receiveMessageResult?: ReceiveMessageResult;
+	private readonly responseMetadata: ResponseMetadata;
+
+	public constructor(parameters: ReceiveMessageResponseInterface)
+	{
+		if (parameters.ReceiveMessageResult !== null)
+		{
+			this.receiveMessageResult = new ReceiveMessageResult(parameters.ReceiveMessageResult);
+		}
+
+		this.responseMetadata = new ResponseMetadata(parameters.ResponseMetadata);
+	}
+
+	public getReceiveMessageResult(): ReceiveMessageResult | undefined
+	{
+		return this.receiveMessageResult;
+	}
+
+	public getResponseMetadata(): ResponseMetadata
+	{
+		return this.responseMetadata;
+	}
+}
+
+export { ReceiveMessageResponse };
