+++ packages/aws-sqs/src/entity/receive-message-result/definition/interface/receive-message-result.interface.mts
@@ -1,0 +1,9 @@
+import type { MessageInterface } from "../../../message/definition/interface/message.interface.mjs";
+
+interface ReceiveMessageResultInterface
+{
+	messages?: Array<MessageInterface> | null;
+	Message?: MessageInterface | null;
+}
+
+export type { ReceiveMessageResultInterface };
