+++ packages/aws-sqs/src/entity/response-envelope/definition/interface/response-envelope.interface.mts
@@ -1,0 +1,8 @@
+import type { ReceiveMessageResponseInterface } from "../../../receive-message-response/definition/interface/receive-message-response.interface.mjs";
+
+interface ResponseEnvelopeInterface
+{
+	ReceiveMessageResponse: ReceiveMessageResponseInterface;
+}
+
+export type { ResponseEnvelopeInterface };
