+++ packages/aws-sqs/src/entity/response-envelope/definition/predicate/assert-response-envelope.mts
@@ -1,0 +1,14 @@
+import type { ResponseEnvelopeInterface } from "../interface/response-envelope.interface.mjs";
+import { assertStructuredData } from "@vitruvius-labs/ts-predicate";
+import { assertReceiveMessageResponse } from "../../../receive-message-response/definition/predicate/assert-receive-message-response.mjs";
+
+function assertResponseEnvelope(value: unknown): asserts value is ResponseEnvelopeInterface
+{
+	assertStructuredData<ResponseEnvelopeInterface>(value, {
+		ReceiveMessageResponse: {
+			test: assertReceiveMessageResponse,
+		},
+	});
+}
+
+export { assertResponseEnvelope };
