+++ packages/aws-sqs/src/entity/response-metadata/definition/interface/response-metadata.interface.mts
