+++ packages/aws-sqs/src/entity/response-metadata/definition/predicate/assert-response-metadata.mts
@@ -1,0 +1,13 @@
+import type { ResponseMetadataInterface } from "../interface/response-metadata.interface.mjs";
+import { assertString, assertStructuredData } from "@vitruvius-labs/ts-predicate";
+
+function assertResponseMetadata(value: unknown): asserts value is ResponseMetadataInterface
+{
+	assertStructuredData<ResponseMetadataInterface>(value, {
+		RequestId: {
+			test: assertString,
+		},
+	});
+}
+
+export { assertResponseMetadata };
