--- packages/aws-sqs/src/predicate/_index.mts
@@ -1,6 +1,0 @@
-export * from "./assert-message-attribute-interface.mjs";
-export * from "./assert-message-interface.mjs";
-export * from "./assert-receive-message-response.mjs";
-export * from "./assert-receive-message-result.mjs";
-export * from "./assert-response-envelope.mjs";
-export * from "./assert-response-metadata.mjs";
