--- packages/aws-sqs/src/service/_index.mts
@@ -1,1 +1,0 @@
-export * from "./sqs.service.mjs";
